backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
python-decouple==3.8
Pillow==9.5.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
whitenoise==6.4.0
python-dotenv==1.0.0
boto3==1.26.0
django-storages==1.13.0
requests==2.31.0
